Okay no debugger on travisci. use ENV['DEBUGGER'] check
right now debugger is breaking on travisci because it's failing on the dependent gem debugger-linecache (1.1.1) which is outdated. gem command can't specify precision of version like Gemfile.lock (which is not checked in for gems), so i can't tell it to install debugger-linecache 1.1.2 or greater instead of 1.1.1. anyway guess removing debugger as default will make more people happier